Fix ssh-dsa -> ssh-dss
Bitvise, FlowSsh and Georgia SSH incorrectly were listed as support
"ssh-dsa" hostkeys instead of "ssh-dss". In each case, this was due to
documentation of the products listing the wrong id in some place.
Luckily, all also list the correct id, so I am reasonably confident that
these are just typos.